requirements.txt
Flask==2.3.2
numpy==1.23.5
pandas==2.0.2
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.5.0
tensorflow==2.16.1
tensorflow_hub==0.16.1